Struct pagerduty::integrations::ResolveEvent

pub struct ResolveEvent<'a> {
    // some fields omitted
}

Cause the referenced incident to enter the resolved state.

Once an incident is resolved, it won't generate any additional notifications. New trigger events with the same incident_key as a resolved incident won't re-open the incident. Instead, a new incident will be created. Your monitoring tools should send PagerDuty a resolve event when the problem that caused the initial trigger event has been fixed.

Methods

impl<'a> ResolveEvent<'a>

fn new<S>(service_key: S, incident_key: S) -> Self where S: Into<Cow<'a, str>>

Create a new ResolveEvent

service_key: The GUID of one of your "Events API" services. This is the "service key" listed on a Generic API's service detail page.

incident_key: Identifies the incident to resolve. This should be the incident_key you received back when the incident was first opened by a trigger event. Resolve events referencing resolved or nonexistent incidents will be discarded.

Examples

// Create a resolve event
let event = ResolveEvent::new("service_key", "incident_key");

fn set_details<T: ?Sized>(self, details: &T) -> Self where T: Serialize

Set event details

An arbitrary JSON object containing any data you'd like included in the incident log.

Examples

// Extra data to be included with the event. Anything that implements
// Serialize can be passed to `set_details`.
#[derive(Serialize)]
struct Details {
    what: &'static str,
    extinguished: i32,
}

// Create a resolve event and include custom data
ResolveEvent::new("service_key", "incident_key")
    .set_details(&Details {
         what: "Server fire",
         extinguished: 1,
    });

fn set_description<S>(self, description: S) -> Self where S: Into<Cow<'a, str>>

Set text that will appear in the incident's log associated with this event.